drivers: pva: support greater than 4GB buffers
add support for buffers greater than 4GB

* cleanup 64bit buffer size when pinning/mapping
* add support for nvpva_pointer_symbol_ex
* update address_range check accordingly

Bug 3814029
Bug 4087600

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2864477
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2904176
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/host/pva/nvpva_buffer.c
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: GPL-2.0 */
 /*
- * Copyright (c) 2016-2022, NVIDIA Corporation.  All rights reserved.
+ * Copyright (c) 2016-2023, NVIDIA Corporation.  All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms and conditions of the GNU General Public License,
@@ -379,7 +379,7 @@ int nvpva_buffer_submit_pin_id(struct nvpva_buffers *nvpva_buffers,
 			       u32 count,
 			       struct dma_buf **dmabuf,
 			       dma_addr_t *paddr,
-			       size_t *psize,
+			       u64 *psize,
 			       enum nvpva_buffers_heap *heap)
 {
 	struct nvpva_vm_buffer *vm;
